why did the commit message change kaleb

+++ src/main/resources/db/migration/procedures/brs/commissions/R__get_closer_residual_detals.sql
@@ -20,8 +20,8 @@ CREATE or replace function brs.get_closer_residual_details(p_closer_user_id bigi
             clawback_projects                  json,
             total_qualifying_fdc_to_date       json,
             cancelled_fdc_during_period        integer,
-            reactivated_fdc                    integer,
-            residual_qualified_fdc             integer,
+            reactivated_fdc                    bigint,
+            residual_qualified_fdc             bigint,
             no_previous_month_message          text
           )
 AS
